audiotest.py changed to now test PyAudio
As far as I'm aware, you'd have to do some complicated shit to get simple audio to play a continuous stream of samples. So, looking for other solutions, PyAudio came along. This code snippet plays a sine wave at 440 hz indefinitely until the console is interrupted (Ctrl-C when run in cmd). Now I'm sure there's a better way to implement this, but this would be a starting point if we get to a real time implementation.

Also for this to work you need to install PyAudio. And for PyAudio to work you need to install PortAudio. pip was having some trouble installing it so download the wheel from this link (https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio) and install it that way if you want to test this out.

Also that is a fuckin pain in the ass, so it's probably better to stick with simpleaudio anyways